Linked add farmer to database
