+++ .flutter-plugins
@@ -1,23 +1,23 @@
 # This is a generated file; do not edit or check into version control.
-cloud_firestore=/home/nishant/flutter/.pub-cache/hosted/pub.dartlang.org/cloud_firestore-3.5.1/
-cloud_firestore_web=/home/nishant/flutter/.pub-cache/hosted/pub.dartlang.org/cloud_firestore_web-2.8.10/
-firebase_auth=/home/nishant/flutter/.pub-cache/hosted/pub.dartlang.org/firebase_auth-3.11.2/
-firebase_auth_web=/home/nishant/flutter/.pub-cache/hosted/pub.dartlang.org/firebase_auth_web-4.6.1/
-firebase_core=/home/nishant/flutter/.pub-cache/hosted/pub.dartlang.org/firebase_core-1.24.0/
-firebase_core_web=/home/nishant/flutter/.pub-cache/hosted/pub.dartlang.org/firebase_core_web-1.7.3/
-firebase_database=/home/nishant/flutter/.pub-cache/hosted/pub.dartlang.org/firebase_database-9.1.7/
-firebase_database_web=/home/nishant/flutter/.pub-cache/hosted/pub.dartlang.org/firebase_database_web-0.2.1+9/
-firebase_messaging=/home/nishant/flutter/.pub-cache/hosted/pub.dartlang.org/firebase_messaging-13.1.0/
-firebase_messaging_web=/home/nishant/flutter/.pub-cache/hosted/pub.dartlang.org/firebase_messaging_web-3.2.0/
-fluttertoast=/home/nishant/flutter/.pub-cache/hosted/pub.dartlang.org/fluttertoast-8.1.1/
-libphonenumber=/home/nishant/flutter/.pub-cache/hosted/pub.dartlang.org/libphonenumber-2.0.2/
-libphonenumber_plugin=/home/nishant/flutter/.pub-cache/hosted/pub.dartlang.org/libphonenumber_plugin-0.2.3/
-libphonenumber_web=/home/nishant/flutter/.pub-cache/hosted/pub.dartlang.org/libphonenumber_web-0.2.0+1/
-path_provider_linux=/home/nishant/flutter/.pub-cache/hosted/pub.dartlang.org/path_provider_linux-2.1.7/
-path_provider_windows=/home/nishant/flutter/.pub-cache/hosted/pub.dartlang.org/path_provider_windows-2.1.3/
-shared_preferences=/home/nishant/flutter/.pub-cache/hosted/pub.dartlang.org/shared_preferences-2.0.17/
-shared_preferences_android=/home/nishant/flutter/.pub-cache/hosted/pub.dartlang.org/shared_preferences_android-2.0.15/
-shared_preferences_foundation=/home/nishant/flutter/.pub-cache/hosted/pub.dartlang.org/shared_preferences_foundation-2.1.3/
-shared_preferences_linux=/home/nishant/flutter/.pub-cache/hosted/pub.dartlang.org/shared_preferences_linux-2.1.3/
-shared_preferences_web=/home/nishant/flutter/.pub-cache/hosted/pub.dartlang.org/shared_preferences_web-2.0.4/
-shared_preferences_windows=/home/nishant/flutter/.pub-cache/hosted/pub.dartlang.org/shared_preferences_windows-2.1.3/
+cloud_firestore=/home/nishant/.pub-cache/hosted/pub.dev/cloud_firestore-3.5.1/
+cloud_firestore_web=/home/nishant/.pub-cache/hosted/pub.dev/cloud_firestore_web-2.8.10/
+firebase_auth=/home/nishant/.pub-cache/hosted/pub.dev/firebase_auth-3.11.2/
+firebase_auth_web=/home/nishant/.pub-cache/hosted/pub.dev/firebase_auth_web-4.6.1/
+firebase_core=/home/nishant/.pub-cache/hosted/pub.dev/firebase_core-1.24.0/
+firebase_core_web=/home/nishant/.pub-cache/hosted/pub.dev/firebase_core_web-1.7.3/
+firebase_database=/home/nishant/.pub-cache/hosted/pub.dev/firebase_database-9.1.7/
+firebase_database_web=/home/nishant/.pub-cache/hosted/pub.dev/firebase_database_web-0.2.1+9/
+firebase_messaging=/home/nishant/.pub-cache/hosted/pub.dev/firebase_messaging-13.1.0/
+firebase_messaging_web=/home/nishant/.pub-cache/hosted/pub.dev/firebase_messaging_web-3.2.0/
+fluttertoast=/home/nishant/.pub-cache/hosted/pub.dev/fluttertoast-8.1.1/
+libphonenumber=/home/nishant/.pub-cache/hosted/pub.dev/libphonenumber-2.0.2/
+libphonenumber_plugin=/home/nishant/.pub-cache/hosted/pub.dev/libphonenumber_plugin-0.2.3/
+libphonenumber_web=/home/nishant/.pub-cache/hosted/pub.dev/libphonenumber_web-0.2.0+1/
+path_provider_linux=/home/nishant/.pub-cache/hosted/pub.dev/path_provider_linux-2.1.7/
+path_provider_windows=/home/nishant/.pub-cache/hosted/pub.dev/path_provider_windows-2.1.3/
+shared_preferences=/home/nishant/.pub-cache/hosted/pub.dev/shared_preferences-2.0.17/
+shared_preferences_android=/home/nishant/.pub-cache/hosted/pub.dev/shared_preferences_android-2.0.15/
+shared_preferences_foundation=/home/nishant/.pub-cache/hosted/pub.dev/shared_preferences_foundation-2.1.3/
+shared_preferences_linux=/home/nishant/.pub-cache/hosted/pub.dev/shared_preferences_linux-2.1.3/
+shared_preferences_web=/home/nishant/.pub-cache/hosted/pub.dev/shared_preferences_web-2.0.4/
+shared_preferences_windows=/home/nishant/.pub-cache/hosted/pub.dev/shared_preferences_windows-2.1.3/
