+++ lib/collectorPages/database_pages/db_add_milk_buyer.dart
@@ -2,7 +2,8 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 
 Future<bool> addNewMilkBuyer(
     String? name, String? phoneNumber, bool morning, bool evening) async {
-  String path = "/Dairy/Dudhaganga/Milk_Buyer";
+  String path = "/Dairy/Dudhaganga/Milk_Buyer/";
+  print("$name $phoneNumber $morning $evening");
   CollectionReference collectionReference =
       FirebaseFirestore.instance.collection(path);
   try {
@@ -14,9 +15,11 @@ Future<bool> addNewMilkBuyer(
         "evening": evening,
       },
     );
+    print("ADDED");
     return true;
   } catch (e) {
     print("Error while adding farmer:$e");
+    print("NOT ADDED");
     return false;
   }
 }
